pv-plant-website/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
pydantic-settings>=2.0.3
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
redis>=5.0.1
python-dotenv>=1.0.0
httpx>=0.25.2
pandas>=2.1.3
numpy>=1.24.4
scipy>=1.11.4
matplotlib>=3.8.2
plotly>=5.17.0
Pillow>=10.1.0
reportlab>=4.0.7
openpyxl>=3.1.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
requests>=2.31.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
pvlib>=0.10.0